Add files via upload
Servomotor is dc motor with controller and some gearbox. It is very helpful to using it in the robotics prototypes. I will discuss 2 types of using it with arduino uno:

First: connect it directly with arduino uno . You can connect it divectly on the arduino uno but olny one motor can uno run it.